Implemented Product and Color controllers

+++ src/Shop/Shop.Domain/ProductAggregate/Value Objects/Score.cs
@@ -5,7 +5,7 @@ namespace Shop.Domain.ProductAggregate.Value_Objects;
 
 public class Score : BaseValueObject
 {
-    public int Value { get; private set; }
+    public float Value { get; private set; }
 
     private const int MinimumScore = 0;
     private const int MaximumScore = 5;
@@ -15,13 +15,13 @@ public class Score : BaseValueObject
 
     }
 
-    public Score(int scoreValue)
+    public Score(float scoreValue)
     {
         Guard(scoreValue);
         Value = scoreValue;
     }
 
-    private void Guard(int scoreValue)
+    private void Guard(float scoreValue)
     {
         OutOfRangeValueDomainException.CheckRange(MinimumScore, MaximumScore, scoreValue, nameof(scoreValue));
     }
